requirements.txt
apiview==1.3.25
celery==5.4.0
eventlet
redis
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.11
django-import-export==3.3.8
djangorestframework==3.15.1
requests==2.31.0
schedule==1.2.1
tencentcloud-sdk-python==3.0.1132
tencentcloud-sdk-python-common==3.0.1132
tencentcloud-sdk-python-hunyuan==3.0.1132
urllib3==2.2.1
wcwidth==0.2.13
xlrd==2.0.1
xlwt==1.3.0
PyJWT==2.8.0
pytz==2024.1
